requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
fastapi==0.109.2
groq==0.18.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
uvicorn==0.27.1
jinja2==3.1.3
pydantic==2.6.1
python-multipart==0.0.9
huggingface-hub==0.16.4
torch==2.0.1
transformers==4.30.2
